implemented some of the changes from class
